+++ app/src/main/java/com/example/android/publishapp/data/repository/IPublishRepository.java
@@ -5,10 +5,13 @@ import com.example.android.publishapp.data.model.PublishModel;
 import java.util.Map;
 
 import io.reactivex.Single;
+import retrofit2.Call;
 
 public interface IPublishRepository {
 
     Single<PublishModel> insertPublishModel(PublishModel publishModel);
 
     Single<Map<String, PublishModel>> getPublishModelList();
+
+    Call<Map<String, PublishModel>> getPublishModelListCallback(int page, int pageSize);
 }
