+++ app/src/main/java/com/example/android/publishapp/di/component/AppComponent.java
@@ -2,6 +2,7 @@ package com.example.android.publishapp.di.component;
 
 import com.example.android.publishapp.di.module.AppModule;
 import com.example.android.publishapp.di.module.FirebaseModule;
+import com.example.android.publishapp.di.module.PagerModule;
 import com.example.android.publishapp.di.module.PublishModule;
 import com.example.android.publishapp.di.module.RetrofitModule;
 import com.example.android.publishapp.presentation.mvp.ui.MainActivity;
@@ -15,7 +16,7 @@ import javax.inject.Singleton;
 import dagger.Component;
 
 @Singleton
-@Component(dependencies = {}, modules = {PublishModule.class, RetrofitModule.class, AppModule.class, FirebaseModule.class})
+@Component(dependencies = {}, modules = {PublishModule.class, PagerModule.class, RetrofitModule.class, AppModule.class, FirebaseModule.class})
 public interface AppComponent {
 
     void inject(MainActivity mainActivity);
