+++ app/src/main/java/com/example/android/publishapp/domain/iteractor/IPublishIteractor.java
@@ -5,10 +5,13 @@ import com.example.android.publishapp.data.model.PublishModel;
 import java.util.Map;
 
 import io.reactivex.Single;
+import retrofit2.Call;
 
 public interface IPublishIteractor {
 
     Single<PublishModel> insertPostInDb(PublishModel publishModel);
 
     Single<Map<String, PublishModel>> getAllPostsFromDb();
+
+    Call<Map<String, PublishModel>> getAllPostsFromDbCallback(int page, int pageSize);
 }
