+++ app/src/main/java/com/example/android/publishapp/domain/iteractor/PublishIteractorImpl.java
@@ -8,6 +8,7 @@ import java.util.Map;
 import javax.inject.Inject;
 
 import io.reactivex.Single;
+import retrofit2.Call;
 
 public class PublishIteractorImpl implements IPublishIteractor {
     private IPublishRepository publishRepository;
@@ -25,5 +26,10 @@ public class PublishIteractorImpl implements IPublishIteractor {
     @Override
     public Single<Map<String, PublishModel>> getAllPostsFromDb() {
         return publishRepository.getPublishModelList();
+    }
+
+    @Override
+    public Call<Map<String, PublishModel>> getAllPostsFromDbCallback(int page, int pageSize) {
+        return publishRepository.getPublishModelListCallback(page, pageSize);
     }
 }
