+++ app/src/main/java/com/example/android/publishapp/presentation/mvp/ui/MainActivity.java
@@ -1,5 +1,7 @@
 package com.example.android.publishapp.presentation.mvp.ui;
 
+import android.arch.lifecycle.LiveData;
+import android.arch.paging.PagedList;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
@@ -14,7 +16,8 @@ import com.arellomobile.mvp.presenter.InjectPresenter;
 import com.arellomobile.mvp.presenter.ProvidePresenter;
 import com.example.android.publishapp.R;
 import com.example.android.publishapp.data.model.PublishModel;
-import com.example.android.publishapp.presentation.adapter.PublishAdapterRv;
+import com.example.android.publishapp.presentation.adapter.PublishDiffUtilCallback;
+import com.example.android.publishapp.presentation.adapter.PublishPagedListAdapter;
 import com.example.android.publishapp.presentation.app.App;
 import com.example.android.publishapp.presentation.mvp.presenter.MainPresenter;
 import com.example.android.publishapp.presentation.mvp.view.MainView;
@@ -32,7 +35,11 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
     RecyclerView recyclerView;
     @BindView(R.id.txt_empty_list)
     TextView textEmptyList;
-    private PublishAdapterRv adapter;
+    private PublishPagedListAdapter pagerAdapter;
+    @Inject
+    LiveData<PagedList<PublishModel>> pagedListLiveData;
+    @Inject
+    PublishDiffUtilCallback publishDiffUtil;
     @Inject
     @InjectPresenter
     MainPresenter mainPresenter;
@@ -48,9 +55,14 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         ButterKnife.bind(this);
-        adapter = new PublishAdapterRv();
+        pagerAdapter = new PublishPagedListAdapter(publishDiffUtil.diffUtilCallback);
         recyclerView.setLayoutManager(new LinearLayoutManager(getApplicationContext(), OrientationHelper.VERTICAL, false));
-        recyclerView.setAdapter(adapter);
+        recyclerView.setAdapter(pagerAdapter);
+        setupPagedList();
+    }
+
+    private void setupPagedList() {
+        pagedListLiveData.observe(this, publishModels -> pagerAdapter.submitList(publishModels));
     }
 
     @Override
@@ -68,7 +80,7 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
     public void setupPublishList(List<PublishModel> publishList) {
         recyclerView.setVisibility(View.VISIBLE);
         textEmptyList.setVisibility(View.GONE);
-        adapter.setupPublishers(publishList);
+      //  adapter.setupPublishers(publishList);
     }
 
     @OnClick(R.id.float_button)
@@ -81,4 +93,5 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
         super.onResume();
         mainPresenter.initPublishers();
     }
+
 }
