+++ app/src/main/res/layout/recycler_event.xml
@@ -13,6 +13,12 @@
         android:layout_height="wrap_content"
         android:orientation="vertical">
 
+        <TextView
+            android:id="@+id/load_txt_event"
+            android:layout_gravity="center_horizontal"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" />
+
         <ViewFlipper
             android:id="@+id/view_flipper_event"
             android:layout_width="@dimen/event_width_picture_size"
