+++ app/src/main/res/layout/recycler_link.xml
@@ -5,6 +5,12 @@
     android:layout_marginBottom="8dp"
     android:orientation="vertical">
 
+    <TextView
+        android:id="@+id/load_txt_link"
+        android:layout_gravity="center_horizontal"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
+
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
