+++ app/src/main/res/values/strings.xml
@@ -25,4 +25,5 @@
     <string name="radio_link">Ссылка</string>
     <string name="enter_date_txt">Дата:*</string>
     <string name="success_post">Пост успешно опубликован</string>
+    <string name="load_txt">Загрузка</string>
 </resources>
